syscalls/fanotify09: Add test case for events with filename info
Kernel v5.9 has a bug when there is a mark on filesystem or mount
interested in events with filename AND there is an additional
mark on a parent watching children.

In some situations the event is reported to filesystem or mount mark
without the filename info.

This bug is fixed by kernel commit 7372e79c9eb9:
  fanotify: fix logic of reporting name info with watched parent

The test case requires a blockdev filesystem.

Reviewed-by: Jan Kara <jack@suse.cz>
Reviewed-by: Petr Vorel <pvorel@suse.cz>
Signed-off-by: Amir Goldstein <amir73il@gmail.com>