shm_open/37-1: Remove non-portable output
This test could mess up some tools, because it wrote characters
outside of the ansi character set. Since most systems expect utf-8 now,
this could lead to messed up or even output.

This replaces the verbatim output of the shm ames used in the test with
a short description.

It also replaces the no-protable characters with escape sequences and
adds the missing close, if the shm was opend successfully.

Signed-off-by: Joerg Vehlow <joerg.vehlow@aox-tech.de>
Reviewed-by: Cyril Hrubis <chrubis@suse.cz>