af_alg07: Add linux-git reference to more general fix
While trying to reintroduce this error on kernel 5.10 I found that a
secondary fix has been applied which fixes many more socket types and
may obsolete the original fix. Ideally we would create a new test
based on Eric's reproducer, but for now I will just include a
reference to the second fix.

Signed-off-by: Richard Palethorpe <rpalethorpe@suse.com>
Cc: Martin Doucha <mdoucha@suse.cz>
Reviewed-by: Petr Vorel <pvorel@suse.cz>