runltp: Generate HTML output file in read-only fs
[ Upstream commit c5e8c14a8df94963b2417eaf23f7be5d76483db3 ]

solves the issue of running LTP from a read-only filesystem. But
generating HTML output file fails. This commit solves the issue of
generating HTML file in read-only filesystem.

Without this patch the script attempted to create a $LTPROOT/output
directory even if HTMLFILE was absulute path, which was, of course,
wrong.

Signed-off-by: Bhargav Das <bhargav_das@mentor.com>
Reviewed-by: Cyril Hrubis <chrubis@suse.cz>