lapi: Move struct file_handle into lapi/fcntl.h
that way it can be used in fanotify tests
(some of use the struct, but not name_to_handle_at() syscall)
and the struct is defined in <fcntl.h> anyway.

Although detection with HAVE_NAME_TO_HANDLE_AT works (at least on glibc,
musl and uclibc-ng) add proper autotools check for the struct presence.

Reviewed-by: Xiao Yang<yangx.jy@cn.fujitsu.com>
Reviewed-by: Cyril Hrubis <chrubis@suse.cz>
Signed-off-by: Petr Vorel <pvorel@suse.cz>