IMA: Add test for SELinux measurement
New functionality has been added in IMA to measure data that is
critical to the integrity of the system.  SELinux uses this feature
to measure the hash of the SELinux policy loaded in kernel memory,
and the current state of various SELinux configuration.

Add test cases which verify that the IMA subsystem correctly
measures the data provided by SELinux.

Relevant kernel commits:
* fdd1ffe8a812 ("selinux: include a consumer of the new IMA critical data hook")
(v5.12-rc1-dontuse)
* 2554a48f4437 ("selinux: measure state and policy capabilities")
(to be in v5.13)

Reviewed-by: Petr Vorel <pvorel@suse.cz>
Signed-off-by: Lakshmi Ramasubramanian <nramas@linux.microsoft.com>