swapping01: check memory swap usage per process
Issue:

Currently test swapping01 read the system FreeSwap for counting usage of
swap-size, that's not precise on system especially with eating-memory daemon
in the background.

Improvement:

Considering that swap out unreferenced pages also could be occurred in a
single process. So that we encapsulate all of the means in a single process
to avoid involving the whole system swap-counting. Then only try to check the
'VmSwap' in proc/PID/status per process, to get rid of the potential influence
from other processes which easily leads to false positive.  e.g.

Child:
         touch unreferenced pages (via alloc&write&free 1.3*MemAvailable) [1]
         alloc&wirte 1.3*MemAvailable
         raise(SIGSTOP)
         ...
Parent:
         waiting for child suspension
         check child's VmSwap to see if heavy-swap occurs in a process
         ...

[1] As to perform alloc&write&free, the system pages will go through a
completed life cycle from buddy-system to active-list to inactive-list
then back to buddy-system, which reflect to a page status is theoretically like:
"inactive,unreferenced -> active,referenced -> ... ->inactive,unreferenced"
so that will helpful to produce what the kernel target commit fixed situation.

New reproducer works well with unfix-kernel:

  # uname  -r
  2.6.39

  # ./swapping01
  tst_test.c:1263: TINFO: Timeout per run is 0h 05m 00s
  swapping01.c:110: TINFO: available physical memory: 1896 MB
  swapping01.c:113: TINFO: try to allocate: 2466 MB
  swapping01.c:148: TFAIL: heavy swapping detected: 1905 MB swapped.

Signed-off-by: Li Wang <liwang@redhat.com>
Reviewed-by: Petr Vorel <pvorel@suse.cz>