pidns13: Convert to new API and fix race condition
Firstly the test used sleep() for synchronisation which is unreliable
and slow, so it has been converted to checkpoints.

Secondly this still leaves a race between wake and sigtimedwait. Also
at least two I/O signals are generated by the pipe FD. One for the
write and probably one for EOF when the other end is closed. So we
block the signal as suggested by the man pages.

Reviewed-by: Petr Vorel <pvorel@suse.cz>
Signed-off-by: Richard Palethorpe <rpalethorpe@suse.com>
Signed-off-by: Petr Vorel <pvorel@suse.cz>