Makefile: Avoid wildcard determinism issues
Wildcard used in Makefiles preserves the order of the files on disk
which leads to differences in the order the binaries are linked.

This patch tweaks the usage of wildcard to also have a sort, making
the binaries reproducible. A previous patch massively improved
the situation but wider testing showed that these other changes can
have an effect on some binaries, this patch avoids those issues as
exposed by our wider testing.

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>
Reviewed-by: Cyril Hrubis <chrubis@suse.cz>