syscalls/swapon/swapoff: Move common library to libs/
There is a race/determinism issue with libswapon currently as depending on
which makefile path builds it, the debug paths may be "../swapon/libswapon.c"
or "libswapon.c". To avoid the determinism issue, move libswapon to libs/
and use it from the shared location.

The header is moved to the common includes directory to match and
the Makefile.inc fragment can be dropped.

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>
Reviewed-by: Cyril Hrubis <chrubis@suse.cz>