IMA: Allow only ima-buf template for key measurement
ima-buf is the default IMA template used for all buffer measurements
since kernel v5.11. Therefore, IMA policy rule for measuring keys need
not specify an IMA template.  But if a template is specified for key
measurement rule it must be only ima-buf.

Update keys tests to not require a template to be specified for
key measurement rule, but if a template is specified verify it is
only ima-buf.

Keep template=ima-buf in keycheck.policy to be compatible with kernel
v5.6 through v5.10.

Reviewed-by: Petr Vorel <pvorel@suse.cz>
Signed-off-by: Lakshmi Ramasubramanian <nramas@linux.microsoft.com>
Signed-off-by: Petr Vorel <pvorel@suse.cz>