realtime/scripts: Fix indentation error in parser.py
An indentation error in parser.py causes certain blocks
of code to run even before dependent variables are
initialized when the code in this file is imported
by parse-testpi1.py or parse-testpi2.py.

Post running tests like testpi-2 or testpi-4, the
results of these tests are parsed and added to the
logs/ directory. This bug prevents the results of such
tests from being parsed and thus masks the results of
these tests.

Fixes: 5a8408bc3 ("*.py : convert python2 to python3")

Reviewed-by: Petr Vorel <pvorel@suse.cz>
Signed-off-by: Sharan Turlapati <sturlapati@vmware.com>