controllers/cpuset: correct type of variable c in checkopt
Test cpuset_load_balance failed on ARM64 bare metal as:

cpuset_load_balance 1 TFAIL: There is something wrong with test tasks
cpuset_load_balance 3 TFAIL: There is something wrong with test tasks
...
cpuset_load_balance 5 TFAIL: There is something wrong with test tasks

The analysis shows that on the ARM64-based machine, the getopt performs
a while operation. In addition, the value of char is 0 instead of 1 when
it is extended to the left during comparison with the integer type,
resulting in a failure at while ((c = getopt(argc, argv, "p:h")) != -1).
Change the type of variable c to int to fix the problem.

Signed-off-by: Wang Xin <wangxin410@huawei.com>
Reviewed-by: Li Wang <liwang@redhat.com>