docparse: Handle special characters in JSON
* escape backslash (\) and double quote (")
  escaping backslash effectively escapes other C escaped strings (\t,
  \n, ...), which we sometimes want (in the comment) but sometimes not
  (in .option we want to have them interpreted)
* replace tab with 8x space
* skip and warn into stderr invalid chars (< 0x20, i.e. anything before space)
  defined by RFC 8259 (https://tools.ietf.org/html/rfc8259#page-9)

NOTE: atm the fix is required only for ". But tab was also problematic in the past.

TODO: This is just a "hot fix" solution before release. For for a
correct decision whether \n, \t should be escaped or interpreted we
should decide in the parser which has the context. C string should be
probably interpreted (thus nothing needed to be done as it escapes in a
compatible way with JSON), but comments probably should display \n, \t
thus add extra \.

Fixes: c39b29f0a ("bpf: Check truncation on 32bit div/mod by zero")

Co-developed-by: Petr Vorel <pvorel@suse.cz>
Signed-off-by: Cyril Hrubis <chrubis@suse.cz>
[ pvorel: commit message ]
Signed-off-by: Petr Vorel <pvorel@suse.cz>