bpf: Add map_array helper functions/macro and constify
We usually just create an array map with standard key and value
size. Then fetch some values from the array. So we can create some
standard functions for this.

The byte code which stores the values in the map can also be
extracted into a macro. So a macro has been added, however it is only
used in the new bpf_prog05 test. The old tests do some pointer
arithmetic on the map call return value to trigger bugs. They could be
rewritten, but then would need verifying.

This also adds some extra const checks and attributes.

Signed-off-by: Richard Palethorpe <rpalethorpe@suse.com>
Reviewed-by: Cyril Hrubis <chrubis@suse.cz>