shmget02: fix EPERM test when RLIMIT_MEMLOCK is large enough
shmget(..., SHM_HUGETLB) will not fail with EPERM if there is enough space
for RLIMIT_MEMLOCK. If the limit is 0, it will fail with EPERM as expected
when a not enough privileged user calls it.

Set RLIMIT_MEMLOCK to 0 for all calls, as this is the only one which will
fail because of that limit, unless we are talking about very old kernels
(before 2.6.9).

Fixes: 4dc493b44a85 ("syscalls/shmget*: Convert into new api")
Signed-off-by: Thadeu Lima de Souza Cascardo <cascardo@canonical.com>
Reviewed-by: Yang Xu <xuyang2018.jy@fujitsu.com