mmap18: fix when PTHREAD_STACK_MIN < PAGE_SIZE
PTHREAD_STACK_MIN can be less than the page size, to be more
precise 2048 on musl. This value is used in the test as an
offset for a new stack address with mmap() + MAP_FIXED flag.
Though it might not be aligned to the page size.

This breaks the test with musl:

  tst_test.c:1311: TINFO: Timeout per run is 0h 05m 00s
  mmap18.c:98: TBROK: mmap(0x7fe67e2ee800,2048,3,306,-1,0) failed: EINVAL (22)
  mmap18.c:169: TFAIL: Child: exited with 2
  [...]

The fix is to align mapped_size arg to the page size.

Signed-off-by: Alexey Kodanev <aleksei.kodanev@bell-sw.com>
Reviewed-by: Cyril Hrubis <chrubis@suse.cz>