syscalls/shmget06: Add test when shm_next_id is already in use
When the identifier of the System V shared memory segment that shm_next_id stored
is already in use, shmget() with different key will return the another shm id. But
kernel doesn't guarantee desired id, I just compare with the existed id, if not
equal, the test succeeds.

This case is similar to msgget05.c.

Reviewed-by: Cyril Hrubis <chrubis@suse.cz>
Signed-off-by: Yang Xu <xuyang2018.jy@fujitsu.com>