ipsec_lib.sh: check ip_vti/ip6_vti are enabled or skip tests
In network stress test groups there are tests expecting
CONFIG_NET_IPVTI/CONFIG_IPV6_VTI to be enabled in the kernel,
and if it's not they fail. There is a check for VTI support in
the ip utility, but not for the kernel. Skip these tests if
there is no ip_vti/ip6_vti support in the kernel.

Reviewed-by: Petr Vorel <pvorel@suse.cz>
Reviewed-by: Alexey Kodanev <aleksei.kodanev@bell-sw.com>
Signed-off-by: Radoslav Kolev <radoslav.kolev@suse.com>
[ simplify check, fix whitespace ]
Signed-off-by: Petr Vorel <pvorel@suse.cz>