syscalls/fanotify22: Test file event with broken inode
This test corrupts an inode entry with an invalid mode through debugfs
and then tries to access it.  This should result in a ext4 error, which
we monitor through the fanotify group.

In order for this test to succeed, it requires a kernel fix introduced
by kernel commit 124e7c61deb2 ("ext4: fix error code saved on super
block during file system abort") from v5.16, which is added to .tags.

Reviewed-by: Amir Goldstein <amir73il@gmail.com>
Reviewed-by: Petr Vorel <pvorel@suse.cz>
Signed-off-by: Gabriel Krisman Bertazi <krisman@collabora.com>