setsockopt08: Avoid confusion by removing TCONF
When not in compat mode this only tests that bad offsets result in
EINVAL. When in compat mode, we also test for the vulnerability.

Therefor I view this as two tests. The TCONF was supposed to indicate
the configuration didn't allow one of these tests. However it just
caused confusion with multiple people asking me to change it to
tst_brk.

So change it to TINFO.

Reported-by: Martin Doucha <mdoucha@suse.cz>
Reviewed-by: Martin Doucha <mdoucha@suse.cz>
Reviewed-by: Petr Vorel <pvorel@suse.cz>
Signed-off-by: Richard Palethorpe <rpalethorpe@suse.com>