lib/tst_test.sh: fix ROD_SILENT command return status check
"local tst_out=$(some_command)" always returns status 0, i.e.
not the return status of the command in the assignment, but the
status for 'local' builtin command, which creates a new local
variable inside the function.

As a result, the library calls and tests that invoke ROD_SILENT
don't actually check the return status of commands passed to it.

The old API library function is also fixed.

Signed-off-by: Alexey Kodanev <aleksei.kodanev@bell-sw.com>
Reviewed-by: Petr Vorel <pvorel@suse.cz>
Tested-by: Petr Vorel <pvorel@suse.cz>